Convert test to TS and import fixtures
